List applicationRoot at the end among all modules
In RuntimeUpdateProcessor:636 the first module to be traversed is always applicationRoot. Whenever a breaking change (e.g, a change in method signature) is made in another module, applicationRoot will always be compiled first. This will result in a compile error. ApplicationRoot should always be the last module to be compiled.